pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-framework"
version = "1.0.0"
description = "Open-source malware analysis sandbox with behavioral monitoring and automated report generation"
authors = [{name = "Sentinel Framework Contributors"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pefile>=2023.2.7",
    "yara-python>=4.3.0",
    "volatility3>=2.4.0",
    "docker>=6.0.0",
    "requests>=2.31.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "colorama>=0.4.6",
    "python-magic>=0.4.27",
    "ssdeep>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sentinel*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
